mmc: mtk-sd: avoid error when second IO resource missing.
mtk-sd allows a second IO resource to be specified which
is stored in "top_base".  This is optional.
However if not provided, an "invalid resource" error is generated.
This is because NULL is passed to devm_ioremap_resource().

So if "res" is NULL, skip that call - top_base will have been
initalized to NULL, so nothing else need be done.

Signed-off-by: NeilBrown <neil@brown.name>